+++ autogpt/prompts/prompt.py
@@ -37,101 +37,3 @@ def build_default_prompt_generator(config: Config) -> PromptGenerator:
         prompt_generator.add_best_practice(best_practice)
 
     return prompt_generator
-
-
-def __print_json_log(config: AIConfig):
-    logger.json_report(
-        "system",
-        """{} has been created with the following details
-    Name: {}
-    Role: {}
-    Goals: {}
-    API Budget: {} 
-        """.format(
-            config.ai_name,
-            config.ai_name,
-            config.ai_role,
-            config.ai_goals,
-            "infinite" if config.api_budget <= 0 else f"${config.api_budget}",
-        ),
-    )
-
-
-def construct_main_ai_config() -> AIConfig:
-    """Construct the prompt for the AI to respond to
-
-    Returns:
-        str: The prompt string
-    """
-    config = AIConfig.load(CFG.ai_settings_file)
-    if CFG.skip_reprompt and config.ai_name:
-        logger.typewriter_log("Name :", Fore.GREEN, config.ai_name)
-        logger.typewriter_log("Role :", Fore.GREEN, config.ai_role)
-        logger.typewriter_log("Goals:", Fore.GREEN, f"{config.ai_goals}")
-        logger.typewriter_log(
-            "API Budget:",
-            Fore.GREEN,
-            "infinite" if config.api_budget <= 0 else f"${config.api_budget}",
-        )
-        __print_json_log(config)
-    elif config.ai_name:
-        logger.typewriter_log(
-            "Welcome back! ",
-            Fore.GREEN,
-            f"Would you like me to return to being {config.ai_name}?",
-            speak_text=True,
-        )
-        logger.json_report(
-            "system",
-            "Welcome back! Would you like me to return to being {}?".format(
-                config.ai_name
-            ),
-        )
-        should_continue = clean_input(
-            logger.json_report(
-                "system",
-                f"""Continue with the last settings?
-    Name:  {config.ai_name}
-    Role:  {config.ai_role}
-    Goals: {config.ai_goals}
-    API Budget: {"infinite" if config.api_budget <= 0 else f"${config.api_budget}"}
-    Continue ({CFG.authorise_key}/{CFG.exit_key}): """,
-            )
-        )
-        if should_continue.lower() == CFG.exit_key:
-            config = AIConfig()
-
-    if not config.ai_name:
-        config = prompt_user()
-        config.save(CFG.ai_settings_file)
-
-    if CFG.restrict_to_workspace:
-        logger.typewriter_log(
-            "NOTE:All files/directories created by this agent can be found inside its workspace at:",
-            Fore.YELLOW,
-            f"{CFG.workspace_path}",
-        )
-    # set the total api budget
-    api_manager = ApiManager()
-    api_manager.set_total_budget(config.api_budget)
-
-    # Agent Created, print message
-    logger.typewriter_log(
-        config.ai_name,
-        Fore.LIGHTBLUE_EX,
-        "has been created with the following details:",
-        speak_text=True,
-    )
-    # Print the ai config details
-    # Name
-    logger.typewriter_log("Name:", Fore.GREEN, config.ai_name, speak_text=False)
-    # Role
-    logger.typewriter_log("Role:", Fore.GREEN, config.ai_role, speak_text=False)
-    # Goals
-    logger.typewriter_log("Goals:", Fore.GREEN, "", speak_text=False)
-    for goal in config.ai_goals:
-        logger.typewriter_log("-", Fore.GREEN, goal, speak_text=False)
-
-    __print_json_log(config)
-
-    return config
